inbound_filter.go: move sendComposedProfile to the control plane
