+++ privacy-profile-composer/pkg/composer/grpc_client.go
@@ -21,6 +21,7 @@ package composer
 import (
 	"context"
 	"flag"
+	"github.com/envoyproxy/envoy/contrib/golang/common/go/api"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -74,4 +75,51 @@ func Run_client() {
 		log.Fatalf("got this error when fetching system wide profile: %v", err)
 	}
 	log.Println(profile)
+}
+
+func sendComposedProfile(fqdn string, purpose string, piiTypes []string, thirdParties []string) api.StatusType {
+	var (
+		composerSvcAddr = flag.String("addr", "http://prose-server.prose-system.svc.cluster.local:50051", "the address to connect to")
+	)
+
+	flag.Parse()
+	// Set up a connection to the server.
+	conn, err := grpc.Dial(*composerSvcAddr, grpc.WithTransportCredentials(insecure.NewCredentials()))
+	if err != nil {
+		log.Printf("can not connect to Composer SVC at addr %v. ERROR: %v", composerSvcAddr, err)
+		return api.Continue
+	}
+	defer func(conn *grpc.ClientConn) {
+		err = conn.Close()
+		if err != nil {
+			log.Printf("could not close connection to Composer server %s", err)
+		}
+	}(conn)
+	c := pb.NewPrivacyProfileComposerClient(conn)
+
+	// Contact the server and print out its response.
+	ctx := context.Background()
+
+	processingEntries := make(map[string]*pb.DataItemAndThirdParties, len(piiTypes))
+	for _, pii := range piiTypes {
+		dataItemThirdParties := map[string]*pb.ThirdParties{
+			pii: {
+				ThirdParty: thirdParties,
+			},
+		}
+		processingEntries[purpose] = &pb.DataItemAndThirdParties{Entry: dataItemThirdParties}
+	}
+	_, err = c.PostObservedProfile(
+		ctx,
+		&pb.SvcObservedProfile{
+			SvcInternalFQDN: fqdn,
+			ObservedProcessingEntries: &pb.PurposeBasedProcessing{
+				ProcessingEntries: processingEntries},
+		},
+	)
+
+	if err != nil {
+		log.Printf("got this error when posting observed profile: %v", err)
+	}
+	return api.Continue
 }
